#39138. 'Undefined passenger pools' bug fixed

+++ src/baggage.cc
@@ -3,7 +3,6 @@
 #include "astra_utils.h"
 #include "base_tables.h"
 #include "term_version.h"
-#include "astra_misc.h"
 #include "qrys.h"
 #include "baggage_tags.h"
 #include <serverlib/cursctl.h>
@@ -1142,8 +1141,41 @@ void TGroupBagItem::toDB(int grp_id) const
   }
 }
 
+void TGroupBagItem::copyPaxPool(const GrpId_t& src, const GrpId_t& dest)
+{
+  if (src==dest) return;
+
+  auto cur=make_curs("MERGE INTO pax "
+                     "USING "
+                     "( "
+                     "SELECT pax1.bag_pool_num, pax2.pax_id "
+                     "FROM pax pax1, tckin_pax_grp tckin_pax_grp1, "
+                     "     pax pax2, tckin_pax_grp tckin_pax_grp2 "
+                     "WHERE pax1.grp_id=tckin_pax_grp1.grp_id AND "
+                     "      pax2.grp_id=tckin_pax_grp2.grp_id AND "
+                     "      tckin_pax_grp1.first_reg_no-pax1.reg_no=tckin_pax_grp2.first_reg_no-pax2.reg_no AND "
+                     "      tckin_pax_grp1.grp_id=:grp_id_src AND "
+                     "      tckin_pax_grp2.grp_id=:grp_id_dest "
+                     ") src "
+                     "ON (pax.pax_id=src.pax_id) "
+                     "WHEN MATCHED THEN "
+                     "UPDATE SET pax.bag_pool_num=src.bag_pool_num, "
+                     "           pax.tid=DECODE(pax.bag_pool_num, src.bag_pool_num, pax.tid, cycle_tid__seq.currval)");
+  cur.bind(":grp_id_src", src.get())
+     .bind(":grp_id_dest", dest.get())
+     .exec();
+}
+
+void TGroupBagItem::copyPaxPool(const std::list<TCkinRouteInsertItem> &tckinGroups)
+{
+  for(const auto& i : tckinGroups)
+    copyPaxPool(tckinGroups.front().grpId, i.grpId);
+}
+
 void TGroupBagItem::copyDB(const GrpId_t& src, const GrpId_t& dest)
 {
+  if (src==dest) return;
+
   auto cur=make_curs("BEGIN "
                      "  DELETE FROM value_bag WHERE grp_id=:grp_id_dest; "
                      "  DELETE FROM unaccomp_bag_info WHERE grp_id=:grp_id_dest; "
@@ -1166,26 +1198,12 @@ void TGroupBagItem::copyDB(const GrpId_t& src, const GrpId_t& dest)
                      "  INSERT INTO unaccomp_bag_info(grp_id,num,original_tag_no,surname,name,airline,flt_no,suffix,scd) "
                      "  SELECT :grp_id_dest,num,original_tag_no,surname,name,airline,flt_no,suffix,scd "
                      "  FROM unaccomp_bag_info WHERE grp_id=:grp_id_src; "
-                     "  MERGE INTO pax "
-                     "  USING "
-                     "  ( "
-                     "  SELECT pax1.bag_pool_num, pax2.pax_id "
-                     "  FROM pax pax1, tckin_pax_grp tckin_pax_grp1, "
-                     "       pax pax2, tckin_pax_grp tckin_pax_grp2 "
-                     "  WHERE pax1.grp_id=tckin_pax_grp1.grp_id AND "
-                     "        pax2.grp_id=tckin_pax_grp2.grp_id AND "
-                     "        tckin_pax_grp1.first_reg_no-pax1.reg_no=tckin_pax_grp2.first_reg_no-pax2.reg_no AND "
-                     "        tckin_pax_grp1.grp_id=:grp_id_src AND "
-                     "        tckin_pax_grp2.grp_id=:grp_id_dest "
-                     "  ) src "
-                     "  ON (pax.pax_id=src.pax_id) "
-                     "  WHEN MATCHED THEN "
-                     "  UPDATE SET pax.bag_pool_num=src.bag_pool_num, "
-                     "             pax.tid=DECODE(pax.bag_pool_num, src.bag_pool_num, pax.tid, cycle_tid__seq.currval); "
                      "END; ");
   cur.bind(":grp_id_src", src.get())
      .bind(":grp_id_dest", dest.get())
      .exec();
+
+  copyPaxPool(src, dest);
 }
 
 void TGroupBagItem::checkTagUniquenessOnFlight(int grp_id)
